+++ src/main/java/pe/edu/unmsm/sistemas/repository/ProgramaRepository.java
@@ -3,6 +3,6 @@ package pe.edu.unmsm.sistemas.repository;
 import org.springframework.data.repository.CrudRepository;
 import pe.edu.unmsm.sistemas.domain.Programa;
 
-public interface ProgramaRepository extends CrudRepository<Programa, Integer> {
+public interface ProgramaRepository extends CrudRepository<Programa, Short> {
 
 }
