+++ src/main/java/pe/edu/unmsm/sistemas/repository/ProgramacionPagoRepository.java
@@ -3,6 +3,6 @@ package pe.edu.unmsm.sistemas.repository;
 import org.springframework.data.repository.CrudRepository;
 import pe.edu.unmsm.sistemas.domain.ProgramacionPago;
 
-public interface ProgramacionPagoRepository extends CrudRepository<ProgramacionPago, Long> {
+public interface ProgramacionPagoRepository extends CrudRepository<ProgramacionPago, Integer> {
 
 }
